refactor: update imports based on refactored chat module in eval
