+++ src/wandbot/evaluation/eval.py
@@ -8,9 +8,9 @@ from fuzzywuzzy import fuzz
 from langchain.chat_models import ChatOpenAI
 from langchain.evaluation.qa import QAEvalChain
 from tqdm.auto import tqdm
-from wandbot.chat import Chat
+from wandbot.chat.chat import Chat
+from wandbot.chat.prompts import load_eval_prompt
 from wandbot.evaluation.config import EvalConfig
-from wandbot.prompts import load_eval_prompt
 
 eval_config = EvalConfig()
